% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_state_tifs.R
\name{process_state_tifs}
\alias{process_state_tifs}
\title{Stack and save downloaded state climate data}
\usage{
process_state_tifs(
  state_name,
  write = FALSE,
  clim_path,
  save_path,
  mask,
  remove = FALSE,
  return = TRUE
)
}
\arguments{
\item{state_name}{Character. State abbreviation as it appears in your file directory.}

\item{write}{Logical. Whether to write the the raster stack as an .RDS file.}

\item{clim_path}{Character. The path of your .tif data.}

\item{save_path}{Character. If writing, where to save to.}

\item{mask}{sf. Converted to SpatVector prior to mask the raster stack.}

\item{remove}{Logical. Whether to remove the individual .tif files used to build the stack after stacking.}

\item{return}{Logical. Return the stack?}
}
\description{
Stack and save downloaded state climate data
}
