% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_extract.R
\name{stack_extract}
\alias{stack_extract}
\title{Extract .tif data to trap points}
\usage{
stack_extract(
  dat,
  state_stack,
  var = c("tmean", "ppt", "tmin", "tmax", "vpdmax", "vpdmin", "tdmean"),
  lon_col = "longitude",
  lat_col = "latitude",
  crs_in = 4326
)
}
\arguments{
\item{dat}{Data frame. Must contain longitude and latitude columns.}

\item{state_stack}{Raster stack containing climate data for a state.}

\item{var}{Character. One of "tmean", "ppt", "tmin", "tmax", "vpdmax", or "tdmean" (variables extractable in prism8_daily).}

\item{lat_col}{Character.}

\item{crs_in}{Numeric. The coordinate reference system to apply to dat.}

\item{lon_ol}{Character.}
}
\description{
Extract .tif data to trap points
}
