% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweeper_fun.R
\name{sweep_fun}
\alias{sweep_fun}
\title{Preprocess mosquito surveillance data}
\usage{
sweep_fun(
  path = NULL,
  state_name,
  extensions = NULL,
  sheets = FALSE,
  type = "abundance"
)
}
\arguments{
\item{path}{Character. The directory of a single datasheet. Defaults to NULL.}

\item{state_name}{Character. Abbreviated state--e.g. "TX", "WA"--that vector data comes from.}

\item{extensions}{If working with multi-year data that is spread across multiple files, a list of path directories. Defaults to NULL.}

\item{sheets}{TRUE or FALSE. Indicates whether data is stored within multiple excel sheets. Defaults to FALSE.}

\item{type}{The type of datasheet to be cleaned: either 'abundance' or 'pool'. Defaults to 'abundance'.}
}
\value{
A preprocessed datasheet.
}
\description{
This functions takes a raw, potentially messy .csv, .xlsx, or .xls file that
contains mosquito abundance of pool data and both cleans and standardizes
species collection and geospatial information.
}
