% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism8_daily.R
\name{prism8_daily}
\alias{prism8_daily}
\title{Download and convert daily PRISM data at 800m resolution to .tif files.}
\usage{
prism8_daily(
  var,
  start_date,
  end_date,
  date_list = NULL,
  dir = getwd(),
  remove = TRUE,
  bil = TRUE,
  template,
  state_name = NULL,
  progress = TRUE
)
}
\arguments{
\item{var}{Character, or list of character. The type of data you would like to download.  Must be "ppt", "tmean", "tmin", "tmax", "tdmean", "vpdmin", or "vpdmax".}

\item{start_date}{Date. The first date in a range of days to download data for.}

\item{end_date}{Date. The last date in a range of days to download data for.}

\item{date_list}{List of date. If you already know which dates you'd like to download data frame, use this to avoid downloading irrelevant data. Preferred method.}

\item{dir}{Character. The location that data will be downloaded to. Defaults to working directory.}

\item{remove}{Logical. Removes zip files after extraction; HIGHLY recommended. Defaults to TRUE.}

\item{bil}{Logical. Removes all extracted files that do not end in .bil or .hdr; which are the minimum files needed to build a raster. Defaults to TRUE.}

\item{template}{bbox. The extent(s) to which to crop downloaded data. Can be a list.}

\item{state_name}{Character. If template is not missing, then the state(s) for which the bounding box is pulled from. Used for writing raster images.}

\item{progress}{Logical. Whether to create a message in the console at the beginning of each file download initialization.}
}
\description{
Download and convert daily PRISM data at 800m resolution to .tif files.
}
\details{
If using lists for template and state_name args, the elements must be aligned to avoid misnamed .tif files.
E.g., if the first element of template is a bounding box for Wyoming then the first element of state_name must be "WY", and so on.
}
\examples{

box1 <- matrix(
 c(-124.56624, 41.99179, -116.46350, 46.29083),
 nrow = 1,
 byrow = TRUE,
 dimnames = list(NULL, c("xmin", "ymin", "xmax", "ymax"))
)

box2 <- matrix(
 c(-106.64565, 25.83738, -93.50829, 36.50070),
 nrow = 1,
 byrow = TRUE,
 dimnames = list(NULL, c("xmin", "ymin", "xmax", "ymax"))
)

boxes <- list(box1, box2)

prism8_daily(var = c("ppt", "tmean", "tmin", "tmax"),
start_date = "2025-12-25",
end_date = "2025-12-26",
template = boxes,
state_name = c("OR", "TX")
)

# alternatively, do not crop data
prism8_daily(var = c("ppt", "tmean", "tmin", "tmax"),
start_date = "2025-12-25",
end_date = "2025-12-26"
)


}
