% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism8_daily.R
\name{prism8_daily}
\alias{prism8_daily}
\title{Extract daily PRISM data at 800m resolution.}
\usage{
prism8_daily(var, start_date, end_date, path = getwd())
}
\arguments{
\item{var}{Character. The type of data you would like to download.  Must be "ppt", "tmean", "tmin", "tmax", "tdmean", "vpdmin", or "vpdmax".}

\item{start_date}{Date. The first date in a range of days to download data for.}

\item{end_date}{Date. The last date in a range of days to download data for.}

\item{path}{Character. The location that data will be downloaded to. Defaults to working directory.}
}
\description{
Extract daily PRISM data at 800m resolution.
}
